requirements.txt
discord-py-interactions
pymongo
python-dateutil
pyyaml